+++ cmd/get/get.go
@@ -1,0 +1,43 @@
+package get
+
+import (
+	"ktrouble/config"
+
+	"github.com/spf13/cobra"
+)
+
+var getCmd = &cobra.Command{
+	Use:   "get",
+	Args:  cobra.MinimumNArgs(1),
+	Short: "Get various internal configuration and kubernetes resource listings",
+	Long: `EXAMPLE:
+  These are mostly utility commands to review things important to running ktrouble.
+  Allowing a display of various items stored in config.yaml and listing various
+  kubernetes resources.
+
+  > ktrouble get namespaces
+  > ktrouble get utilities
+
+EXAMPLE:
+  Get a list of PODs that are currently running on the current context kubernetes
+  cluster that were created with the ktrouble utility.  If the 'enableBashLinks'
+  config.yaml setting is 'true', a '<bash: ... >' command will be displayed,
+  otherwise the SHELL path will be displayed.
+
+  > ktrouble get pods
+
+    NAME                NAMESPACE       STATUS   EXEC
+    basic-tools-e1df2f  common-tooling  Running  <bash:kubectl -n common-tooling exec -it basic-tools-e1df2f -- /bin/bash>
+
+    NAME                NAMESPACE       STATUS   SHELL
+    basic-tools-e1df2f  common-tooling  Running  /bin/bash
+`,
+	Run: func(cmd *cobra.Command, args []string) {},
+}
+
+var c *config.Config
+
+func InitSubCommands(conf *config.Config) *cobra.Command {
+	c = conf
+	return getCmd
+}
