+++ cmd/root.go
@@ -2,28 +2,22 @@ package cmd
 
 import (
 	"fmt"
-	"math/rand"
 	"os"
+	"regexp"
 	"strings"
-	"time"
 
+	"ktrouble/cmd/get"
 	"ktrouble/common"
 	"ktrouble/config"
+	"ktrouble/defaults"
+	"ktrouble/kubernetes"
 	"ktrouble/objects"
-
-	homedir "github.com/mitchellh/go-homedir"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
 
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	// "gopkg.in/yaml.v3"
 	// "sigs.k8s.io/yaml"
-)
-
-const (
-	uniqIdLength = 6
 )
 
 type (
@@ -42,27 +36,28 @@ var (
 	gitRef    string
 	buildDate string
 
-	c        = &config.Config{}
-	utilMap  map[string]objects.UtilityPod
-	utilDefs []objects.UtilityPod
-	sizeMap  map[string]objects.ResourceSize
-	sizeDefs []objects.ResourceSize
+	// semVerReg - gets the semVer portion only, cutting off any other release details
+	semVerReg = regexp.MustCompile(`(v[0-9]+\.[0-9]+\.[0-9]+).*`)
+
+	c = &config.Config{}
 )
 
-// rootCmd represents the base command when called without any subcommands
-var rootCmd = &cobra.Command{
+// RootCmd represents the base command when called without any subcommands
+var RootCmd = &cobra.Command{
 	Use:   "ktrouble",
-	Short: "",
+	Short: "A tool for launching PODs from a curated list of utility PODs",
 	Long: `EXAMPLE:
+  Simply run the 'launch' command and you will be prompted for all of the
+  required details.
+    - Utility Pod Selection
+    - Namespace
+    - Service Account
+    - Node Selector
+    - Resource Sizing
 
-  TODO: add description
-
-  > ktrouble
-
+  > ktrouble launch
 `,
 	PersistentPreRun: func(cmd *cobra.Command, args []string) {
-
-		rand.Seed(time.Now().UnixNano())
 
 		logFile, _ := cmd.Flags().GetString("log-file")
 		logLevel, _ := cmd.Flags().GetString("log-level")
@@ -88,7 +83,6 @@ var rootCmd = &cobra.Command{
 		c.VersionDetail.BuildDate = buildDate
 		c.VersionDetail.GitCommit = gitCommit
 		c.VersionDetail.GitRef = gitRef
-		c.EnableBashLinks = viper.GetBool("enableBashLinks")
 		c.VersionJSON = fmt.Sprintf("{\"SemVer\": \"%s\", \"BuildDate\": \"%s\", \"GitCommit\": \"%s\", \"GitRef\": \"%s\"}", semVer, buildDate, gitCommit, gitRef)
 		if c.OutputFormat != "" {
 			c.FormatOverridden = true
@@ -103,23 +97,48 @@ var rootCmd = &cobra.Command{
 			}
 		}
 
-		// if os.Args[1] != "version" { // && os.Args[1] != "config" {
-		// }
+		if os.Args[1] != "version" {
+			c.Client = kubernetes.New()
+			if c.Client == nil {
+				common.Logger.Fatal("failed to create a kubernetes context")
+			}
+		}
 	},
+}
+
+func buildRootCmd() *cobra.Command {
+	RootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.splicectl/config.yml)")
+	RootCmd.PersistentFlags().StringVarP(&c.OutputFormat, "output", "o", "", "output types: json, text, yaml, gron, raw")
+	RootCmd.PersistentFlags().BoolVar(&c.NoHeaders, "no-headers", false, "Suppress header output in Text output")
+	// RootCmd.PersistentFlags().BoolVar(&c., "no-headers", false, "Suppress header output in Text output")
+	RootCmd.PersistentFlags().StringVarP(&c.LogLevel, "log-level", "v", "", "Set the logging level: trace,debug,info,warning,error,fatal")
+	RootCmd.PersistentFlags().StringVar(&c.LogFile, "log-file", "", "Set the logging level: trace,debug,info,warning,error,fatal")
+	RootCmd.PersistentFlags().StringVarP(&c.Namespace, "namespace", "n", "", "Specify the namespace to run in, ENV NAMESPACE then -n for preference")
+
+	return RootCmd
+}
+
+func addSubCommands() {
+	RootCmd.AddCommand(
+		// from 'import ktrouble/cmd/<subcommand:package>'
+		// <package>.InitSubCommands(c),
+		get.InitSubCommands(c),
+	)
 }
 
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute() {
-	cobra.CheckErr(rootCmd.Execute())
+	if err := RootCmd.Execute(); err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
 }
 
 func init() {
+	buildRootCmd()
 	cobra.OnInitialize(initConfig)
-
-	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.ktrouble.yaml)")
-	rootCmd.PersistentFlags().StringVarP(&c.OutputFormat, "output", "o", "", "Set an output format: json, text, yaml, gron, md")
-	rootCmd.PersistentFlags().StringVarP(&c.Namespace, "namespace", "n", "", "Specify the namespace to run in, ENV NAMESPACE then -n for preference")
+	addSubCommands()
 }
 
 // initConfig reads in config file and ENV variables if set.
@@ -159,44 +178,84 @@ func initConfig() {
 	}
 
 	// Utility Definitions
-	err := viper.UnmarshalKey("utilityDefinitions", &utilDefs)
+	err := viper.UnmarshalKey("utilityDefinitions", &c.UtilDefs)
 	if err != nil {
 		logrus.Fatal("Error unmarshalling utility defs...")
 	}
-	if len(utilDefs) == 0 {
+	if len(c.UtilDefs) == 0 {
 		logrus.Warn("Adding default utility definitions to config.yaml")
-		seedDefs := defaultUtilityDefinitions()
+		seedDefs := defaults.UtilityDefinitions()
 		viper.Set("utilityDefinitions", seedDefs)
 		verr := viper.WriteConfig()
 		if verr != nil {
 			logrus.WithError(verr).Info("Failed to write config")
 		}
 	} else {
-		utilMap = make(map[string]objects.UtilityPod, len(utilDefs))
-		for _, v := range utilDefs {
-			utilMap[v.Name] = v
+		c.UtilMap = make(map[string]objects.UtilityPod, len(c.UtilDefs))
+		for _, v := range c.UtilDefs {
+			c.UtilMap[v.Name] = v
 		}
 	}
 
 	// Size Definitions
-	serr := viper.UnmarshalKey("resourceSizing", &sizeDefs)
+	serr := viper.UnmarshalKey("resourceSizing", &c.SizeDefs)
 	if serr != nil {
 		logrus.Fatal("Error unmarshalling resource sizing...")
 	}
-	if len(sizeDefs) == 0 {
+	if len(c.SizeDefs) == 0 {
 		logrus.Warn("Adding default resource sizing to config.yaml")
-		seedSizes := defaultResourceSizingList()
+		seedSizes := defaults.ResourceSizingList()
 		viper.Set("resourceSizing", seedSizes)
 		verr := viper.WriteConfig()
 		if verr != nil {
 			logrus.WithError(verr).Info("Failed to write config")
 		}
 	} else {
-		sizeMap = make(map[string]objects.ResourceSize, len(sizeDefs))
-		for _, v := range sizeDefs {
-			sizeMap[v.Name] = v
+		c.SizeMap = make(map[string]objects.ResourceSize, len(c.SizeDefs))
+		for _, v := range c.SizeDefs {
+			c.SizeMap[v.Name] = v
 		}
 	}
+
+	// Node Selector Labels
+	nerr := viper.UnmarshalKey("nodeSelectorLabels", &c.NodeSelectorLabels)
+	if nerr != nil {
+		logrus.Fatal("Error unmarshalling node selector labels...")
+	}
+	if len(c.NodeSelectorLabels) == 0 {
+		logrus.Warn("Adding default node selector labels to config.yaml")
+		seedLabels := defaults.Labels()
+		viper.Set("nodeSelectorLabels", seedLabels)
+		verr := viper.WriteConfig()
+		if verr != nil {
+			logrus.WithError(verr).Info("Failed to write config")
+		}
+	}
+
+	// Unique ID Length
+	if viper.IsSet("uniqIdLength") {
+		c.UniqIdLength = viper.GetInt("uniqIdLength")
+	} else {
+		// Set the default
+		viper.Set("uniqIdLength", 6)
+		verr := viper.WriteConfig()
+		if verr != nil {
+			logrus.WithError(verr).Info("Failed to write config")
+		}
+	}
+
+	// EnableBashLinks
+	if viper.IsSet("enableBashLinks") {
+		c.EnableBashLinks = viper.GetBool("enableBashLinks")
+	} else {
+		// Set the default
+		viper.Set("enableBashLinks", false)
+		verr := viper.WriteConfig()
+		if verr != nil {
+			logrus.WithError(verr).Info("Failed to write config")
+		}
+	}
+
 }
 
 func createRestrictedConfigFile(fileName string) {
@@ -212,172 +271,18 @@ func createRestrictedConfigFile(fileName string) {
 				logrus.Info("Chmod for config file failed, please set the mode to 0600.")
 			}
 		}
-		viper.SetConfigFile(fileName)
-		defaultLabels := defaultLabelList()
-		logrus.Warn("Writing default labels to config.yaml...")
-		viper.Set("nodeSelectorLabels", defaultLabels)
-
-		defaultSizes := defaultResourceSizingList()
-		logrus.Warn("Writing default sizes to config.yaml...")
-		viper.Set("resourceSizing", defaultSizes)
-
-		verr := viper.WriteConfig()
-		if verr != nil {
-			logrus.WithError(verr).Info("Failed to write config")
-		}
 	}
 }
 
-func Commands() []*cobra.Command {
-	return rootCmd.Commands()
-}
-
-func defaultUtilityDefinitions() []objects.UtilityPod {
-
-	return []objects.UtilityPod{
-		{
-			Name:        "dnsutils",
-			Repository:  "gcr.io/kubernetes-e2e-test-images/dnsutils:1.3",
-			ExecCommand: "/bin/sh",
-		},
-		{
-			Name:        "psql-curl",
-			Repository:  "barrypiccinni/psql-curl:latest",
-			ExecCommand: "/bin/bash",
-		},
-		{
-			Name:        "psqlutils15",
-			Repository:  "postgres:15-bullseye",
-			ExecCommand: "/bin/bash",
-		},
-		{
-			Name:        "psqlutils14",
-			Repository:  "postgres:14-bullseye",
-			ExecCommand: "/bin/bash",
-		},
-		{
-			Name:        "awscli",
-			Repository:  "amazon/aws-cli:latest",
-			ExecCommand: "/bin/bash",
-		},
-		{
-			Name:        "gcloudutils",
-			Repository:  "google/cloud-sdk:latest",
-			ExecCommand: "/bin/bash",
-		},
-		{
-			Name:        "azutils",
-			Repository:  "mcr.microsoft.com/azure-cli",
-			ExecCommand: "/bin/bash",
-		},
-		{
-			Name:        "mysqlutils5",
-			Repository:  "mysql:5.7.40-debian",
-			ExecCommand: "/bin/bash",
-		},
-		{
-			Name:        "mysqlutils8",
-			Repository:  "mysql:8-debian",
-			ExecCommand: "/bin/bash",
-		},
-		{
-			Name:        "redis6",
-			Repository:  "cmaahs/redis-cli:6.2",
-			ExecCommand: "/bin/bash",
-		},
-		{
-			Name:        "curl",
-			Repository:  "curlimages/curl:latest",
-			ExecCommand: "/bin/sh",
-		},
-		{
-			Name:        "basic-tools",
-			Repository:  "cmaahs/basic-tools:v0.0.1",
-			ExecCommand: "/bin/bash",
-		},
-		{
-			Name:        "golang19-bullseye",
-			Repository:  "golang:1.19-bullseye",
-			ExecCommand: "/bin/bash",
-		},
+// ClientSemVer - returns the full semVer as the first string and the numerical
+// portion as the second string, they may be identical. One example where they
+// would not be is:
+//
+//	semVer: v0.1.1-cacert -> (v0.1.1-cacert, v0.1.1).
+func ClientSemVer() (string, string) {
+	submatches := semVerReg.FindStringSubmatch(semVer)
+	if submatches == nil || len(submatches) < 2 {
+		logrus.Fatalf("the semver in the current build is not valid: %s", semVer)
 	}
-
-}
-
-func defaultResourceSizingList() []objects.ResourceSize {
-
-	return []objects.ResourceSize{
-		{
-			Name:       "Small",
-			LimitsCPU:  "250m",
-			LimitsMEM:  "2Gi",
-			RequestCPU: "100m",
-			RequestMEM: "512Mi",
-		},
-		{
-			Name:       "Medium",
-			LimitsCPU:  "500m",
-			LimitsMEM:  "4Gi",
-			RequestCPU: "200m",
-			RequestMEM: "1Gi",
-		},
-		{
-			Name:       "Large",
-			LimitsCPU:  "1000m",
-			LimitsMEM:  "8Gi",
-			RequestCPU: "500m",
-			RequestMEM: "2Gi",
-		},
-		{
-			Name:       "XLarge",
-			LimitsCPU:  "10000m",
-			LimitsMEM:  "30Gi",
-			RequestCPU: "6000m",
-			RequestMEM: "2Gi",
-		},
-	}
-}
-
-func defaultLabelList() []string {
-	return []string{
-		"kubernetes.io/arch",
-		"kubernetes.io/os",
-		"node.kubernetes.io/instance-type",
-		"node_pool",
-	}
-}
-
-func restConfig() (*rest.Config, error) {
-	// We aren't likely to run this INSIDE the K8s cluster, this routine
-	// simply picks up the config from the file system of a running POD.
-	// kubeCfg, err := rest.InClusterConfig()
-	var kubeCfg *rest.Config
-	var err error
-
-	if kubeconfig := os.Getenv("KUBECONFIG"); kubeconfig != "" {
-		kubeCfg, err = clientcmd.BuildConfigFromFlags("", kubeconfig)
-		if err != nil {
-			logrus.Info("No KUBECONFIG ENV")
-			return nil, err
-		}
-	} else {
-		// ENV KUBECONFIG not set, check for ~/.kube/config
-		home, err := homedir.Dir()
-		if err != nil {
-			fmt.Println(err)
-			os.Exit(1)
-		}
-		kubeFile := fmt.Sprintf("%s/%s", home, ".kube/config")
-		if _, err := os.Stat(kubeFile); err != nil {
-			if os.IsNotExist(err) {
-				if os.Args[1] != "version" {
-					logrus.Info("Could not locate the KUBECONFIG file, normally ~/.kube/config")
-					os.Exit(1)
-				}
-				return nil, nil
-			}
-		}
-		kubeCfg, err = clientcmd.BuildConfigFromFlags("", kubeFile)
-	}
-	return kubeCfg, nil
+	return submatches[0], submatches[1]
 }
