+++ cmd/version.go
@@ -2,12 +2,11 @@ package cmd
 
 import (
 	"encoding/json"
-	"fmt"
-	"strings"
 
-	"ktrouble/common"
 	"ktrouble/objects"
 
+	"github.com/pkg/errors"
+	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
@@ -16,39 +15,27 @@ var versionCmd = &cobra.Command{
 	Use:   "version",
 	Short: "Express the 'version' of ktrouble.",
 	Run: func(cmd *cobra.Command, args []string) {
-
-		var verData objects.Version
-		err := json.Unmarshal([]byte(c.VersionJSON), &verData)
+		version, err := expressVersion()
 		if err != nil {
-			common.Logger.WithError(err).Error("Failed to unmarshal JSON")
+			logrus.WithError(err).Error("Failed to express the version")
 		}
-
 		if !c.FormatOverridden {
-			c.OutputFormat = "raw"
+			c.OutputFormat = "json"
 		}
-		fmt.Println(versionDataToString(verData))
-
+		c.OutputData(&version, objects.TextOptions{ NoHeaders: c.NoHeaders })
 	},
 }
 
-func versionDataToString(versionData objects.Version) string {
-
-	switch strings.ToLower(c.OutputFormat) {
-	case "raw":
-		return c.VersionJSON
-	case "json":
-		return versionData.ToJSON()
-	case "gron":
-		return versionData.ToGRON()
-	case "yaml":
-		return versionData.ToYAML()
-	case "text", "table":
-		return versionData.ToTEXT(c.NoHeaders)
-	default:
-		return versionData.ToTEXT(c.NoHeaders)
+func expressVersion() (objects.Version, error) {
+	var verData objects.Version
+	err := json.Unmarshal([]byte(c.VersionJSON), &verData)
+	if err != nil {
+		return verData, errors.Wrap(err, "Failed to unmarshal JSON")
 	}
+
+	return verData, nil
 }
 
 func init() {
-	rootCmd.AddCommand(versionCmd)
+	RootCmd.AddCommand(versionCmd)
 }
