+++ config/config.go
@@ -2,33 +2,45 @@ package config
 
 import (
 	"fmt"
-	"strings"
-
+	"ktrouble/kubernetes"
 	"ktrouble/objects"
+	"strings"
 )
 
 type (
 	Config struct {
-		VersionDetail    objects.Version
-		VersionJSON      string
-		OutputFormat     string
-		FormatOverridden bool
-		NoHeaders        bool
-		CACert           string
-		CABundle         string
-		Namespace        string
-		EnableBashLinks  bool
+		VersionDetail      objects.Version
+		VersionJSON        string
+		OutputFormat       string
+		FormatOverridden   bool
+		NoHeaders          bool
+		CACert             string
+		CABundle           string
+		LogLevel           string
+		LogFile            string
+		Namespace          string
+		EnableBashLinks    bool
+		UniqIdLength       int
+		UtilMap            map[string]objects.UtilityPod
+		UtilDefs           objects.UtilityPodList
+		SizeMap            map[string]objects.ResourceSize
+		SizeDefs           objects.ResourceSizeList
+		NodeSelectorLabels []string
+		Client             kubernetes.KubernetesClient
 	}
+
 	Outputtable interface {
 		ToJSON() string
 		ToYAML() string
 		ToGRON() string
-		ToTEXT(noHeaders bool) string
+		ToTEXT(to objects.TextOptions) string
 	}
 )
 
-func (c *Config) outputData(data Outputtable) string {
+func (c *Config) outputData(data Outputtable, to objects.TextOptions) string {
 	switch strings.ToLower(c.OutputFormat) {
+	case "raw":
+		return fmt.Sprintf("%#v", data)
 	case "json":
 		return data.ToJSON()
 	case "gron":
@@ -36,16 +48,12 @@ func (c *Config) outputData(data Outputtable) string {
 	case "yaml":
 		return data.ToYAML()
 	case "text", "table":
-		return data.ToTEXT(c.NoHeaders)
+		return data.ToTEXT(to)
 	default:
-		return ""
+		return data.ToTEXT(to)
 	}
 }
 
-func (c *Config) OutputData(data Outputtable) {
-	if !c.FormatOverridden {
-		c.OutputFormat = "text"
-	}
-
-	fmt.Println(c.outputData(data))
+func (c *Config) OutputData(data Outputtable, to objects.TextOptions) {
+	fmt.Println(c.outputData(data, to))
 }
