+++ objects/text_options.go
@@ -1,0 +1,10 @@
+package objects
+
+type (
+	TextOptions struct {
+		NoHeaders    bool
+		ShowExec     bool
+		UtilMap      map[string]UtilityPod
+		UniqIdLength int
+	}
+)
