+++ objects/version.go
@@ -5,10 +5,9 @@ import (
 	"encoding/json"
 	"strings"
 
-	"ktrouble/common"
-
 	"github.com/maahsome/gron"
 	"github.com/olekukonko/tablewriter"
+	"github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
 )
 
@@ -23,7 +22,7 @@ type Version struct {
 func (v *Version) ToJSON() string {
 	versionJSON, err := json.MarshalIndent(v, "", "  ")
 	if err != nil {
-		common.Logger.WithError(err).Error("Error extracting JSON")
+		logrus.WithError(err).Error("Error extracting JSON")
 		return ""
 	}
 	return string(versionJSON[:])
@@ -32,14 +31,14 @@ func (v *Version) ToJSON() string {
 func (v *Version) ToGRON() string {
 	versionJSON, err := json.MarshalIndent(v, "", "  ")
 	if err != nil {
-		common.Logger.WithError(err).Error("Error extracting JSON for GRON")
+		logrus.WithError(err).Error("Error extracting JSON for GRON")
 	}
 	subReader := strings.NewReader(string(versionJSON[:]))
 	subValues := &bytes.Buffer{}
 	ges := gron.NewGron(subReader, subValues)
 	ges.SetMonochrome(false)
 	if serr := ges.ToGron(); serr != nil {
-		common.Logger.WithError(serr).Error("Problem generating GRON syntax")
+		logrus.WithError(serr).Error("Problem generating GRON syntax")
 		return ""
 	}
 	return subValues.String()
@@ -48,13 +47,16 @@ func (v *Version) ToGRON() string {
 func (v *Version) ToYAML() string {
 	versionYAML, err := yaml.Marshal(v)
 	if err != nil {
-		common.Logger.WithError(err).Error("Error extracting YAML")
+		logrus.WithError(err).Error("Error extracting YAML")
 		return ""
 	}
 	return string(versionYAML[:])
 }
 
-func (v *Version) ToTEXT(noHeaders bool) string {
+func (v *Version) ToTEXT(to TextOptions) string {
+
+	noHeaders := to.NoHeaders
+
 	buf := new(bytes.Buffer)
 	var row []string
 
@@ -66,7 +68,7 @@ func (v *Version) ToTEXT(noHeaders bool) string {
 	}
 
 	table.SetAutoWrapText(false)
-	table.SetAutoFormatHeaders(false)
+	table.SetAutoFormatHeaders(true)
 	table.SetAlignment(tablewriter.ALIGN_LEFT)
 	table.SetCenterSeparator("")
 	table.SetColumnSeparator("")
